Shorten code
This will shorten the code and change the second if statement to an elif statement and make it part of the first if statement. I also shortened the end by using the walrus operator to set the error value within the return statement. Please keep in mind that this can only be used on Python 3.8 and above if implemented. If this is not in the best interests of the repository, please feel free to discard my pull request.