Implement Walrus Operator
Note: This will only work on Python versions 3.8 and above. If you do not intend on using this code on those versions, please discard this pull request.
This will do the same thing as before, just in a shorter fashion. The walrus operator allows you to set values inside of expressions.
Example:
print(n := 5)
^ This would print 5 to the console window and set n equal to it